event router debug

+++ routers/category_router/crud.py
@@ -77,7 +77,8 @@ async def update_category(id: int, payload: schemas.UpdateCategory, db: Session)
 async def delete_category(id: int, db: Session):
     try:
         category = await read_category_by_id(id, db)
-        if category:
+        if category and len(category.images):
+            await utils.delete_folder(category_DIR+"/"+category.images[0].folder_name)
             db.delete(category)
         db.commit()
         return True
